requirements.txt
BeautifulSoup==3.2.1
Django==1.7.1
Markdown==2.3.1
MySQL-python==1.2.5
Pillow==2.4.0
PyJWT==1.4.0
PyLEMS==0.4.5
South==1.0
Werkzeug==0.11.1
Whoosh==2.5.7
argparse==1.2.1
bbcode==1.0.16
boto==2.28.0
cffi==1.3.0
cryptography==1.1
defusedxml==0.4.1
django-annoying==0.8.0
django-classy-tags==0.5.1
django-debug-toolbar==1.2.2
django-djconfig==0.1.4
django-extensions==1.4.9
django-filter==0.9.2
django-haystack==2.1.0
django-infinite-scroll-pagination==0.1.1
django-mptt==0.6.1
django-sekizai==0.7
django-sslserver==0.15
django-storages==1.1.8
djangorestframework==3.2.3
djangorestframework-xml==1.0.1
enum34==1.0.4
generateDS==2.14a
httplib2==0.9
idna==2.0
ipaddress==1.0.15
lxml==3.4.1
oauth2==1.5.211
oauthlib==0.6.3
pyOpenSSL==0.15.1
pyasn1==0.1.9
pycparser==2.14
pydot==1.0.2
pyparsing==1.5.7
python-magic==0.4.6
python-openid==2.2.5
python-social-auth==0.2.13
pytz==2014.3
requests==2.7.0
requests-oauthlib==0.4.1
six==1.6.1
sorl-thumbnail==11.12.1b
sqlparse==0.1.13
urllib3==1.9.1
wheel==0.24.0
wsgiref==0.1.2
yolk==0.4.3
